backend/requirements.txt
Flask
tweepy
vaderSentiment
python-dotenv
requests
gunicorn